Suggestion to include the imported Class
Pycharm has consistently removed the import of SpacyCore when optimising imports because it is not used explicitly in the code. This little change fixes that and clarifies where the nlp pipe name comes from. Gentle suggestion. 

(Aside, I've started exploring `hmrb` and really like the ease of constructing the grammar objects so far -- thanks for the package!)